altering stats
Changing nStakeCombineAge to 10 days. This will combine inputs that are older than 10 days.
Changing nCombineThreshold to 250 coins. This is a max value for the combination.

What this does is combine coins that have more than 10 days of weight. Weight is defined by how long coins are sitting in a wallet. Inputs can find their own stakes/blocks. but if they dont they will be combined.